getting screenshots when loaded
