+++ private/jsx/store/reducers/datasets.js
@@ -120,13 +120,13 @@ export default (state = initialDatasetsState, action) => {
                     layerKey: l.layerKey,
                 }
             })
-
             return {
                 ...state,
                 layers: _.keyBy(layers, 'layerKey'),
                 minMax: layers[0].geojson.minMax,
                 allIndices: layers[0].allIndices,
                 bounds: layers[0].bounds,
+                getScreenshot: datasets[0].screenshots,
             }
         }
         default:
